Toggle EXTCOMIN pin via pwm
It seems like not toggling it may cause a DC offset in the display
hardware and thus 1) additional power draw and 2) damage to the device!

This is currently handled externally from the module. Not yet sure how
to abstract this away from the hardware. Maybe a pwm trait or something.