Manage twi ressources dynamically
As it turns out, two i2c interfaces are not enough to allocate them
statically in the app. For example, the hrm app will need 1) touch 2)
hrm and 3) accel, all of which need a i2c peripheral. However, we now
just let the different drivers request one of the two peripherals
(waiting if none are available) whenever they actually want to talk on
the interface and release it as soon as they are done. Importantly, they
are not held while waiting for an interrupt.